% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{my_filter}
\alias{my_filter}
\title{My Filter}
\usage{
my_filter(dat)
}
\arguments{
\item{dat}{a data.frame}
}
\value{
a data.frame
}
\description{
A wrapper around dplyr::filter to show NSE with a package.
}
\examples{

dat1 <- dplyr::data_frame(a = c(1, 2, 1, 1, 2, 1),
                          b = c(1, 1, 1, 2, 2, 2),
                          c = 1:6)

a <- c(0, 1, 0, 0, 0, 1)
b <- c(2, 2, 2, 0, 2, 2)

all.equal(dplyr::filter(dat1, a == 1, b == 2),
          my_filter(dat1))

}
