% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary.R
\name{my_filter}
\alias{my_filter}
\title{My Summary}
\usage{
my_filter(x, args)
}
\arguments{
\item{x}{a data.frame}

\item{args}{the filtering conditions}
}
\value{
a data.frame
}
\description{
A simple function for summarizing a data set.
}
\examples{

v1 <- dplyr::filter(mtcars, mpg < 14)
v2 <- my_filter(mtcars, mpg < 14)

if (!all.equal(v1, v2)) {
  stop("my_filter is broken", call. = FALSE)
}

v3 <- dplyr::filter(mtcars, mpg < 14 & hp > 220)
v4 <- my_filter(mtcars, mpg < 14 & hp > 220)

if (!all.equal(v3, v4)) {
  stop("my_filter is broken", call. = FALSE)
}

}
